[FIX]  Pass learning rate by data member

+++ include/gradientDescentOptimizer.h
@@ -2,6 +2,7 @@
 #define GRADIENT_DESCENT
 
 #include <deque>
+#include <unordered_map>
 
 // forward declaration
 template <typename T>
@@ -10,15 +11,17 @@ class Minimizer;
 class GradientDescentOptimizer
 {
 public:
-    GradientDescentOptimizer(int lr);
+    GradientDescentOptimizer(float lr);
     void computeGradients(BaseNode *loss);
 
     template <typename T>
     Minimizer<T> minimize(BaseNode *loss);
 
 private:
-    int learningRate_;
+    float learningRate_;
     std::vector<BaseNode *> variableNodesList_;
 };
+
+#include "../src/gradientDescentOptimizer.tpp"
 
 #endif /* GRADIENT_DESCENT */
